+++ example/lib/main.dart
@@ -1,109 +1,69 @@
-import 'dart:math';
-
 import 'package:bonfire/bonfire.dart';
-import 'package:example/enemy/goblin.dart';
-import 'package:example/interface/knight_interface.dart';
-import 'package:example/map/dungeon_map.dart';
-import 'package:example/player/knight.dart';
-import 'package:flame/animation.dart' as FlameAnimation;
+import 'package:example/game_manual_map.dart';
+import 'package:example/game_tiled_map.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  await Flame.util.setLandscape();
-  await Flame.util.fullScreen();
+  await Flame.util.setLandscape(); //TODO Comment when running for web
+  await Flame.util.fullScreen(); //TODO Comment when running for web
   runApp(
     MaterialApp(
-      home: Game(),
+      home: Menu(),
     ),
   );
 }
 
-class Game extends StatelessWidget implements GameListener {
-  final GameController _controller = GameController();
-
+class Menu extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return BonfireWidget(
-      joystick: Joystick(
-        directional: JoystickDirectional(
-          spriteBackgroundDirectional: Sprite('joystick_background.png'),
-          spriteKnobDirectional: Sprite('joystick_knob.png'),
-          size: 100,
-          isFixed: false,
+    return Scaffold(
+      backgroundColor: Colors.cyan[900],
+      body: Center(
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            Text(
+              'Bonfire',
+              style: TextStyle(fontSize: 30, color: Colors.white),
+            ),
+            SizedBox(
+              height: 30,
+            ),
+            SizedBox(
+              width: 200,
+              child: RaisedButton(
+                shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(20)),
+                child: Text('Manual Map'),
+                onPressed: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(builder: (context) => GameManualMap()),
+                  );
+                },
+              ),
+            ),
+            SizedBox(
+              height: 10,
+            ),
+            SizedBox(
+              width: 200,
+              child: RaisedButton(
+                shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(20)),
+                child: Text('Tiled Map'),
+                onPressed: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(builder: (context) => GameTiledMap()),
+                  );
+                },
+              ),
+            ),
+          ],
         ),
-        actions: [
-          JoystickAction(
-            actionId: 0,
-            sprite: Sprite('joystick_atack.png'),
-            align: JoystickActionAlign.BOTTOM_RIGHT,
-            size: 80,
-            margin: EdgeInsets.only(bottom: 50, right: 50),
-          ),
-          JoystickAction(
-            actionId: 1,
-            sprite: Sprite('joystick_atack_range.png'),
-            spriteBackgroundDirection: Sprite('joystick_background.png'),
-            enableDirection: true,
-            size: 50,
-            margin: EdgeInsets.only(bottom: 50, right: 160),
-          )
-        ],
-      ),
-      player: Knight(
-        Position(5 * DungeonMap.tileSize, 6 * DungeonMap.tileSize),
-      ),
-      interface: KnightInterface(),
-      map: DungeonMap.map(),
-      decorations: DungeonMap.decorations(),
-      enemies: DungeonMap.enemies(),
-      background: BackgroundColorGame(Colors.blueGrey[900]),
-      gameController: _controller..setListener(this),
-      lightingColorGame: Colors.black.withOpacity(0.5),
-      showFPS: true,
-      zoom: 1, // you can change this to change the game zoom or directly on the camera
-    );
-  }
-
-  @override
-  void updateGame() {}
-
-  @override
-  void changeCountLiveEnemies(int count) {
-    if (count < 2) {
-      _addEnemyInWorld();
-    }
-  }
-
-  void _addEnemyInWorld() {
-    double x = DungeonMap.tileSize * (4 + Random().nextInt(25));
-    double y = DungeonMap.tileSize * (5 + Random().nextInt(3));
-
-    Position position = Position(
-      x,
-      y,
-    );
-    _controller.addComponent(
-      AnimatedObjectOnce(
-        animation: FlameAnimation.Animation.sequenced(
-          "smoke_explosin.png",
-          6,
-          textureWidth: 16,
-          textureHeight: 16,
-        ),
-        position: Rect.fromLTWH(
-          position.x,
-          position.y,
-          DungeonMap.tileSize,
-          DungeonMap.tileSize,
-        ),
-      ),
-    );
-
-    _controller.addEnemy(
-      Goblin(
-        position,
       ),
     );
   }