+++ lib/rpg_game.dart
@@ -14,6 +14,7 @@ import 'package:bonfire/util/lighting/lighting_config.dart';
 import 'package:bonfire/util/lighting/with_lighting.dart';
 import 'package:bonfire/util/map_explorer.dart';
 import 'package:bonfire/util/value_generator.dart';
+import 'package:flame/components/component.dart';
 import 'package:flame/keyboard.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -61,21 +62,21 @@ class RPGGame extends BaseGamePointerDetector with KeyboardEvents {
     this.collisionAreaColor,
     this.lightingColorGame,
     double zoom,
-  })  : assert(map != null),
-        assert(context != null),
+  })  : assert(context != null),
         assert(joystickController != null) {
     if (zoom != null) gameCamera = Camera(zoom: zoom);
     gameCamera.gameRef = this;
     joystickController.joystickListener = player ?? MapExplorer(gameCamera);
     if (gameController != null) gameController.setGame(this);
-    if (background != null) add(background);
-    add(map);
-    decorations?.forEach((decoration) => add(decoration));
-    enemies?.forEach((enemy) => add(enemy));
-    if (player != null) add(player);
-    if (lightingColorGame != null) add(Lighting(color: lightingColorGame));
-    if (interface != null) add(interface);
-    add(joystickController);
+    if (background != null) super.add(background);
+    if (map != null) super.add(map);
+    decorations?.forEach((decoration) => super.add(decoration));
+    enemies?.forEach((enemy) => super.add(enemy));
+    if (player != null) super.add(player);
+    if (lightingColorGame != null)
+      super.add(Lighting(color: lightingColorGame));
+    if (interface != null) super.add(interface);
+    super.add(joystickController);
     _interval = IntervalTick(200, tick: _updateTempList);
   }
 
@@ -85,12 +86,13 @@ class RPGGame extends BaseGamePointerDetector with KeyboardEvents {
     super.update(t);
   }
 
-  void addEnemy(Enemy enemy) {
-    add(enemy);
+  void addGameComponent(GameComponent component) {
+    addLater(component);
   }
 
-  void addDecoration(GameDecoration decoration) {
-    add(decoration);
+  @override
+  void add(Component c) {
+    addLater(c);
   }
 
   Iterable<Enemy> visibleEnemies() => _visibleEnemies;