+++ lib/util/lighting/lighting.dart
@@ -2,6 +2,7 @@ import 'dart:ui';
 
 import 'package:bonfire/util/game_component.dart';
 import 'package:bonfire/util/lighting/lighting_config.dart';
+import 'package:bonfire/util/priority_layer.dart';
 import 'package:flutter/material.dart';
 
 class Lighting extends GameComponent {
@@ -18,7 +19,7 @@ class Lighting extends GameComponent {
   }
 
   @override
-  int priority() => 19;
+  int priority() => PriorityLayer.LIGHTING;
 
   @override
   void render(Canvas canvas) {